DestinationCell & SectionHeader moved to their own respective files
